requirements.txt
requests==2.31.0
pandas==2.1.4
beautifulsoup4==4.12.2
boto3==1.34.14
pyarrow==14.0.1
python-dotenv==1.0.0